requirements.txt
matplotlib==3.5.1
numpy==1.22.3
opencv_python==4.5.5.64
Pillow==9.1.0
streamlit==1.8.1
tensorflow==2.8.0
